Parse ruby error response body and add attributes
This commit updates the ApiError class in the generated Ruby gem. The
class has been updated to not pass hash arguments directly to the StandardError
constructor. Only a possible title or message hash attribute is passed
to StandardError. This prevents Ruby from stringifying a hash and using
that as the error message, which is both ugly and unusable.